requirements.txt
# Dependencies
fastapi
beanie
python-jose[cryptography]
python-decouple
passlib
bcrypt
pydantic-settings